general azimuthal averaging
