+++ sarracen/disc/__init__.py
@@ -1,1 +1,2 @@
-from ..disc.surface_density import surface_density, angular_momentum, scale_height, honH
+from ..disc.surface_density import azimuthal_average, surface_density
+from ..disc.surface_density import angular_momentum, scale_height, honH
